refactor: Moved Header and Container component to layouts folder
