+++ resources/js/components/layouts/Container.vue
